#spacer doesn't seem to be recognized in Evo, and i'm attempting to see if preloading it into the gem will help the email partials help the evo partials... It's not clear where #spacer isn't defined. in the error I'm getting on Evo when attempting to use %spacer{ spacer: spacer(:quadruple) }
